+++ changelog.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-> :warning: **Project is currently under development**: Version numbers and changelog starts with first release 
-
-## [0.0.0] - 2024-12-23
+## [UNRELEASED] - 2025-00-00
 ### Added
+- Copy value via keybind in log detail view 
+- Add existing value to filter via keybind in log detail view
 
 ### Changed
+- Log detail page now uses interactive table
 
 ### Removed